Move DTOs to "dto" package.
Not sure why they were in the "domain" package before...
