+++ server/src/test/java/com/adobe/testing/s3mock/dto/ListBucketResultV2Test.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017-2021 Adobe.
+ *  Copyright 2017-2022 Adobe.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@ package com.adobe.testing.s3mock.dto;
 
 import static com.adobe.testing.s3mock.dto.DtoTestUtil.serializeAndAssert;
 
-import com.adobe.testing.s3mock.domain.BucketContents;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
